Break the circular dependency of runtime/runtime on testline by moving 'get_runtime_and_trie_from_genesis' into 'integration-tests'
