+++ integration-tests/Cargo.toml
@@ -41,7 +41,7 @@ near-test-contracts = { path = "../runtime/near-test-contracts" }
 near-vm-errors = { path = "../runtime/near-vm-errors" }
 near-vm-runner = { path = "../runtime/near-vm-runner" }
 nearcore = { path = "../nearcore" }
-node-runtime = { path = "../runtime/runtime"}
+node-runtime = { path = "../runtime/runtime" }
 testlib = { path = "../test-utils/testlib" }
 
 [features]
@@ -50,7 +50,11 @@ regression_tests = []
 expensive_tests = []
 adversarial = ["nearcore/adversarial"]
 metric_recorder = ["near-client-primitives/metric_recorder"]
-protocol_feature_alt_bn128 = ["nearcore/protocol_feature_alt_bn128"]
+protocol_feature_alt_bn128 = [
+    "near-primitives/protocol_feature_alt_bn128",
+    "node-runtime/protocol_feature_alt_bn128",
+    "near-vm-errors/protocol_feature_alt_bn128",
+]
 protocol_feature_block_header_v3 = ["near-primitives/protocol_feature_block_header_v3", "near-chain/protocol_feature_block_header_v3", "near-store/protocol_feature_block_header_v3"]
 nightly_protocol_features = ["nearcore/nightly_protocol_features", "protocol_feature_alt_bn128", "protocol_feature_block_header_v3"]
 nightly_protocol = ["nearcore/nightly_protocol"]
